+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 #### Changed
 
 - Increased application heap memory size
+- Remove properties from geometry before sending to Insights API
+- Don't send feature collection with `null` geometry to Insights API
 
 #### Removed